// packages/vsb-website/types/contentful/section-type.ts
export enum SECTION_TYPE {
  BANNER = 'BANNER',
  MESSAGE = 'MESSAGE',
  TILE_LINKS = 'TILE_LINKS',
  TILE_LINK = 'TILE_LINK',
  CLINIC_PREVIEW = 'CLINIC_PREVIEW',
  CONTACT_PREVIEW = 'CONTACT_PREVIEW',
  CLINIC_MISSION = 'CLINIC_MISSION',
  CLINIC_DR_PROFILE = 'CLINIC_DR_PROFILE',
  VASECTOMY_STEPS = 'VASECTOMY_STEPS',
  VASECTOMY_INFO = 'VASECTOMY_INFO',
  VASECTOMY_INFO_BEFORE = 'VASECTOMY_INFO_BEFORE',
  VASECTOMY_INFO_COST = 'VASECTOMY_INFO_COST',
  VASECTOMY_INFO_AFTER = 'VASECTOMY_INFO_AFTER',
  VASECTOMY_FORM_LINK = 'VASECTOMY_FORM_LINK',
  VASECTOMY_FORM_VIDEO = 'VASECTOMY_FORM_VIDEO',
  VASECTOMY_FORM_FORM = 'VASECTOMY_FORM_FORM',
  PILOT_EXAM_INFO = 'PILOT_EXAM_INFO',
  PILOT_EXAM_SERVICE = 'PILOT_EXAM_SERVICE',
  PILOT_EXAM_DR_PROFILE = 'PILOT_EXAM_DR_PROFILE',
  CONTACT_CONTACT = 'CONTACT_CONTACT',
}